Add root_dir parameter to apt module
It is now possible to specify an alternate root directory when using the
apt module by passing the "root_dir" parameter to the apt module.  This
can be very handy for managing packages on a chroot or NFS exported root
filesystem that is different/distinct from the host's root filesystem.